+++ src/libmach/libmach.h
@@ -36,8 +36,8 @@ struct objops {
 	int (*strip)(Obj *obj);
 	int (*pc2line)(Obj *, unsigned long long , char *, int *);
 
-	Symbol *(*getsym)(Obj *obj, long *index, Symbol *sym);
-	Section *(*getsec)(Obj *obj, long *index, Section *sec);
+	Symbol *(*getsym)(Obj *obj, int *index, Symbol *sym);
+	Section *(*getsec)(Obj *obj, int *index, Section *sec);
 
 	int (*setidx)(long nsyms, char *names[], long offset[], FILE *fp);
 	int (*getidx)(long *nsyms, char ***names, long **offset, FILE *fp);
